# Conflicts:
#	library/foundation/src/vscf_simple_swu.c
#	library/foundation/src/vscf_simple_swu_defs.h
#	library/phe/src/vsce_phe_hash.c
#	library/phe/src/vsce_phe_hash_defs.h
#	library/ratchet/src/vscr_ratchet_defs.h
#	library/ratchet/src/vscr_ratchet_group_participant.c
#	tests/foundation/CMakeLists.txt

+++ library/ratchet/src/vscr_ratchet_group_participant_epoch.h
@@ -85,7 +85,7 @@ struct vscr_ratchet_group_participant_epoch_t {
     //
     VSCR_ATOMIC size_t refcnt;
 
-    size_t epoch;
+    uint32_t epoch;
 
     vscr_ratchet_chain_key_t *chain_key;
